telegram_bot/requirements.txt
paho-mqtt==1.6.1
python-telegram-bot==20.7
